agro_alchemy_backend/app.py
import os
from flask import Flask, request, jsonify
from flask_cors import CORS
import random
from google import genai
from PIL import Image
import io
import pickle
import numpy as np
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Load the ML model
ml_model = None
try:
    with open('pest_fertilizers.pkl', 'rb') as f:
        ml_model = pickle.load(f)
    print("Successfully loaded ML model for fertilizer/crop prediction.")
except Exception as e:
    print(f"Failed to load ML model: {e}")

app = Flask(__name__)
# Enable CORS for all domains to allow Flutter web to connect
CORS(app)

# Configure Gemini API using the new google-genai client
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
if GEMINI_API_KEY:
    client = genai.Client(api_key=GEMINI_API_KEY)
    MODEL_ID = "gemini-2.0-flash"
else:
    print("Warning: GEMINI_API_KEY not found in environment variables.")

def handle_genai_error(e):
    error_msg = str(e)
    if "429" in error_msg or "RESOURCE_EXHAUSTED" in error_msg:
        return jsonify({
            "error": "API Quota Exceeded. Please wait a minute and try again.",
            "is_quota_error": True,
            "fallback": True
        }), 429
    return jsonify({"error": error_msg}), 500

@app.route('/')
def home():
    return "AgroAlchemy Backend is Running!"

# --- Yield Prediction Endpoint ---
@app.route('/predict/yield', methods=['POST'])
def predict_yield():
    data = request.json
    print(f"Received yield request: {data}")
    
    crop = data.get('crop', 'Unknown')
    area = float(data.get('area', 1.0))
    if area == 0:
        area = 1.0
    rainfall = float(data.get('rainfall', 0.0))
    season = data.get('season', 'Unknown')
    state = data.get('state', 'Pune')
    production = float(data.get('production', 0.0))
    
    # Calculate yield deterministically
    yield_ph = production / area
    
    # We will use the ML model (pest_fertilizers.pkl) to predict a fertilizer/pesticide
    recommended_pesticide = "No specific recommendation"
    
    if ml_model is not None:
        try:
            def safe_transform(encoder, value):
                if value in encoder.classes_:
                    return encoder.transform([value])[0]
                for c in encoder.classes_:
                    if str(c).lower() == str(value).lower():
                        return encoder.transform([c])[0]
                return 0
            
            # Use state as district if possible
            district_enc = safe_transform(ml_model['label_encoders']['District_Name'], state)
            soil_enc = safe_transform(ml_model['label_encoders']['Soil_color'], 'Black') # Default soil
            crop_enc = safe_transform(ml_model['crop_encoder'], crop)
            
            # Default missing values: N, P, K, pH, Temp
            # Pass unscaled features since the model was trained on them
            import pandas as pd
            feature_cols = ['District_Name', 'Soil_color', 'Nitrogen', 'Phosphorus', 'Potassium', 'pH', 'Rainfall', 'Temperature']
            fert_features_df = pd.DataFrame(
                [[district_enc, soil_enc, 50.0, 50.0, 50.0, 6.5, rainfall, 25.0]], 
                columns=feature_cols
            )
            fert_features_df['Crop_encoded'] = [crop_enc]
            
            fert_pred = ml_model['fertilizer_model'].predict(fert_features_df)
            recommended_fertilizer = ml_model['fertilizer_encoder'].inverse_transform(fert_pred)[0]
            
            recommended_pesticide = f"{recommended_fertilizer} (ML Prediction)"
            
        except Exception as e:
            print(f"Error predicting with ML model in yield endpoint: {e}")
            
    return jsonify({
        "yield_tons_per_hectare": round(yield_ph, 2),
        "total_yield": round(yield_ph * float(area), 2),
        "pesticide_units": "Kg/Ha",
        "recommended_pesticide": recommended_pesticide
    })

# --- Fertilizer Prediction Endpoint ---
@app.route('/predict/fertilizer', methods=['POST'])
def predict_fertilizer():
    data = request.json
    print(f"Received fertilizer request: {data}")
    
    district = data.get('district_name', 'Unknown')
    soil_color = data.get('soil_color', 'Unknown')
    crop = data.get('crop', 'Unknown')
    n = float(data.get('nitrogen', 0))
    p = float(data.get('phosphorus', 0))
    k = float(data.get('potassium', 0))
    ph = float(data.get('ph', 7.0))
    rainfall = float(data.get('rainfall', 0))
    temp = float(data.get('temperature', 25))
    
    if ml_model is not None:
        try:
            # Map encoders safely
            def safe_transform(encoder, value):
                if value in encoder.classes_:
                    return encoder.transform([value])[0]
                for c in encoder.classes_:
                    if str(c).lower() == str(value).lower():
                        return encoder.transform([c])[0]
                return 0 # Default if label unseen
            
            district_enc = safe_transform(ml_model['label_encoders']['District_Name'], district)
            soil_enc = safe_transform(ml_model['label_encoders']['Soil_color'], soil_color)
            
            import pandas as pd
            feature_cols = ['District_Name', 'Soil_color', 'Nitrogen', 'Phosphorus', 'Potassium', 'pH', 'Rainfall', 'Temperature']
            features_df = pd.DataFrame([[district_enc, soil_enc, n, p, k, ph, rainfall, temp]], columns=feature_cols)
            
            # Predict crop
            crop_pred = ml_model['crop_model'].predict(features_df)
            recommended_crop = ml_model['crop_encoder'].inverse_transform(crop_pred)[0]
            
            # Use user-provided crop if available, otherwise fallback to recommended crop
            use_crop = crop if (crop and crop.lower() != 'unknown') else recommended_crop
            crop_enc = safe_transform(ml_model['crop_encoder'], use_crop)
            
            # Predict fertilizer using unscaled features
            fert_features_df = features_df.copy()
            fert_features_df['Crop_encoded'] = [crop_enc]
            fert_pred = ml_model['fertilizer_model'].predict(fert_features_df)
            recommended_fertilizer = ml_model['fertilizer_encoder'].inverse_transform(fert_pred)[0]
            
            note_crop = recommended_crop if (crop and crop.lower() != 'unknown') else use_crop
            note = f"ML Prediction based on soil and weather data. Recommended additional crop: {note_crop}."
            return jsonify({
                "fertilizer": recommended_fertilizer,
                "note": note,
                "crop_predicted": recommended_crop
            })
            
        except Exception as e:
            print(f"Error in ML model prediction: {e}")
            # Fall back to Gemini API
    
    prompt = f"""
    As an agricultural expert, recommend a fertilizer based on:
    Location: {district}
    Soil Color/Type: {soil_color}
    Crop to be grown: {crop}
    Soil Nutrients: Nitrogen={n}, Phosphorus={p}, Potassium={k}
    pH Level: {ph}
    Environment: Rainfall={rainfall}mm, Temp={temp}°C
    
    Provide the most suitable fertilizer recommendation and a brief note.
    Respond strictly in JSON format with keys: "fertilizer" and "note".
    Example: {{"fertilizer": "Urea + DAP (50kg each)", "note": "Apply after first rain"}}
    """
    
    try:
        response = client.models.generate_content(
            model=MODEL_ID,
            contents=prompt
        )
        import json
        text = response.text.replace('```json', '').replace('```', '').strip()
        result = json.loads(text)
        
        return jsonify({
            "fertilizer": result.get('fertilizer', "NPK 19:19:19"),
            "note": result.get('note', "Apply as per local guidelines.")
        })
    except Exception as e:
        print(f"Error in fertilizer prediction: {e}")
        return handle_genai_error(e)

# --- Disease Prediction Endpoint ---
@app.route('/predict/disease', methods=['POST'])
def predict_disease():
    print("Received disease prediction request")
    
    if 'image' not in request.files:
        return jsonify({"error": "No image uploaded"}), 400
        
    file = request.files['image']
    img_bytes = file.read()
    img = Image.open(io.BytesIO(img_bytes))
    
    prompt = """
    Analyze this plant leaf image and identify:
    1. The name of the disease (or "Healthy" if no disease).
    2. Recommended treatment (fungicide, insecticide, or practice).
    3. Severity level (Low, Medium, High, or None).
    
    Respond strictly in JSON format with keys: "disease", "treatment", "severity".
    Example: {{"disease": "Leaf Blight", "treatment": "Copper Oxychloride 50WP", "severity": "High"}}
    """
    
    try:
        response = client.models.generate_content(
            model=MODEL_ID,
            contents=[prompt, img]
        )
        import json
        text = response.text.replace('```json', '').replace('```', '').strip()
        result = json.loads(text)
        
        return jsonify({
            "disease": result.get('disease', "Unknown"),
            "treatment": result.get('treatment', "Consult expert"),
            "severity": result.get('severity', "Medium"),
            "accuracy": f"{random.randint(90, 98)}%"
        })
    except Exception as e:
        print(f"Error in disease prediction: {e}")
        return handle_genai_error(e)

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)


