agro_alchemy_backend/requirements.txt
flask
flask-cors
passlib
google-genai
python-dotenv
Pillow
